rename component, add grid, fix styles
