+++ src/Destinations/index.js
@@ -1,7 +1,8 @@
 import React from "react";
-import { Container, CardGroupRow, Row, Column, Title } from "../commons/styled";
+import { Grid, Row, Col } from "react-flexbox-grid";
+import { Slider, Slide, SectionTitle } from "../commons/styled";
 import NextLink from "../commons/nextLink";
-import Card from "./cardComponent";
+import Card from "./Card";
 
 import paris from "./paris.png";
 import miami from "./miami.png";
@@ -12,33 +13,35 @@ import la from "./la.png";
 
 export default function() {
   return (
-    <Container>
-      <Row justify="flex-start" align="center">
-        <Column size={12}>
-          <Title>Feature destinations</Title>
-        </Column>
-      </Row>
-      <CardGroupRow>
-        <Column size={2}>
-          <Card title="Paris" img={paris} alt="Paris" />
-        </Column>
-        <Column size={2}>
-          <Card title="Miami" img={miami} alt="Miami" />
-        </Column>
-        <Column size={2}>
-          <Card title="Tokyo" img={tokyo} alt="Tokyo" />
-        </Column>
-        <Column size={2}>
-          <Card title="Cape Town" img={capetown} alt="Cape Town" />
-        </Column>
-        <Column size={2}>
-          <Card title="Seoul" img={seoul} alt="Seoul" />
-        </Column>
-        <Column size={2}>
-          <Card title="Los Angeles" img={la} alt="Los Angeles" />
-        </Column>
-        <NextLink top={109} />
-      </CardGroupRow>
-    </Container>
+    <section>
+      <Grid>
+        <Row middle="xs">
+          <Col xs>
+            <SectionTitle>Feature destinations</SectionTitle>
+          </Col>
+        </Row>
+        <Slider>
+          <Slide xs={4} md={3} lg={2}>
+            <Card title="Paris" img={paris} alt="Paris" />
+          </Slide>
+          <Slide xs={4} md={3} lg={2}>
+            <Card title="Miami" img={miami} alt="Miami" />
+          </Slide>
+          <Slide xs={4} md={3} lg={2}>
+            <Card title="Tokyo" img={tokyo} alt="Tokyo" />
+          </Slide>
+          <Slide xs={4} md={3} lg={2}>
+            <Card title="Cape Town" img={capetown} alt="Cape Town" />
+          </Slide>
+          <Slide xs={4} md={3} lg={2}>
+            <Card title="Seoul" img={seoul} alt="Seoul" />
+          </Slide>
+          <Slide xs={4} md={3} lg={2}>
+            <Card title="Los Angeles" img={la} alt="Los Angeles" />
+          </Slide>
+          <NextLink top={40} />
+        </Slider>
+      </Grid>
+    </section>
   );
 }
